refactor(cli): compact command moves to src/cli/compact.ts
Pure move of the `compact` command out of main(). The command body,
options, hidden --hook option and hook dispatch path are byte-identical;
only module specifiers change with the new location.
